requirements.txt
flask 
pymongo
dnspython 
python-dotenv